Add to main

+++ internal/akuma-no-mi/model.go
@@ -1,7 +1,7 @@
 package akumanomi
 
 type AkumaNoMi struct {
-	ID          int64  `json:"id" gorm:"primaryKey"`
+	ID          uint   `json:"id" gorm:"primaryKey"`
 	Name        string `json:"name"`
 	Model       string `json:"model"`
 	Meaning     string `json:"meaning"`
